# Optimise HtmlEncode and HtmlDecode to produce less garbage