This fixes https://github.com/dotnet/corefx/issues/13893. I'd noticed in a project of mine that the `TextWriter` overload of `WebUtility.HtmlEncode` was unexpectedly generating garbage.

I added a `System.Buffers` project reference to the `System.Runtime.Extensions` csproj and wrote a simple implementation of `StringBuilder` based on `ArrayPool`, as suggested by @jkotas in https://github.com/dotnet/corefx/pull/13875.

I benchmarked this implementation and while it appears to produce no garbage (except for the `string` that it returns) it does seem to run around 10-20% slower. I'm keen to get some input from you guys on what might be going on here.


Benchmark code (using BenchmarkDotNet):

```csharp
[MemoryDiagnoser]
public class Bench
{
    public static readonly string _toEncode = string.Concat(Enumerable.Repeat("<>\"'", 100));

    [Benchmark(Baseline = true)]
    public void String_Old()
    {
        WebUtility.HtmlEncode(_toEncode);
    }

    [Benchmark]
    public void String_New()
    {
        WebUtility.HtmlEncodeNew(_toEncode);
    }

    [Benchmark]
    public void TextWriter_Old()
    {
        using (var writer = TextWriter.Null)
        {
            WebUtility.HtmlEncode(_toEncode, writer);
        }
    }

    [Benchmark]
    public void TextWriter_New()
    {
        using (var writer = TextWriter.Null)
        {
            WebUtility.HtmlEncodeNew(_toEncode, writer);
        }
    }
}
```

And the results:

```
         Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |  Gen 0 |  Gen 1 | Allocated |
--------------- |---------:|----------:|----------:|-------:|---------:|-------:|-------:|----------:|
     String_Old | 5.394 us | 0.0987 us | 0.0924 us |   1.00 |     0.00 | 2.0828 | 0.0229 |   13120 B |
     String_New | 6.571 us | 0.1101 us | 0.1030 us |   1.22 |     0.03 | 0.7629 |      - |    4832 B |
 TextWriter_Old | 5.431 us | 0.1078 us | 0.1153 us |   1.01 |     0.03 | 2.0828 | 0.0229 |   13120 B |
 TextWriter_New | 6.080 us | 0.1041 us | 0.0974 us |   1.13 |     0.03 |      - |      - |       0 B |
```