@davidfowl 2018-02-19
FYI there's a much more efficient HtmlEncoder implementation here https://github.com/dotnet/corefx/blob/01fa16ffd618846a913cad719269c8bb441ceb28/src/System.Text.Encodings.Web/src/System/Text/Encodings/Web/HtmlEncoder.cs.