@benjamin-hodgson 2018-02-19
@davidfowl Thanks, I didn't know about `HtmlEncoder`.