@benjamin-hodgson 2018-03-07
Rebased. Apologies for the delay in getting this back up to speed.

The tests were failing because I changed the code to throw `ArgumentNullException` when the input string was null, per [@danmosemsft's suggestion](https://github.com/dotnet/corefx/pull/27250#discussion_r169736385). Previously `HtmlEncode`/`HtmlDecode` would return `null` (or write nothing to the `TextWriter`, depending on which overload you called) if the input was `null`. For now [I've updated the tests](https://github.com/dotnet/corefx/pull/27250/commits/36aeacdfdf3ff736456fd161a5eeb0f4d0af55d0) to match the new behaviour, but this is technically a breaking change; it's not unlikely that client code is relying on being able to pass `null` to that method. Do you think I should keep the new behaviour or put it back to how it was?