@ahsonkhan 2018-03-09
@benjamin-hodgson, the System.Web.Tests.HttpUtilityTest/HtmlEncode_TextWriter(decoded: null, encoded: \"\") test is failing in CI 

https://mc.dot.net/#/user/benjamin-hodgson/pr~2Fjenkins~2Fdotnet~2Fcorefx~2Fmaster~2F/test~2Ffunctional~2Fcli~2F/127d2ab60cced49647237cd9c6774f5102fb1095/workItem/System.Web.HttpUtility.Tests/analysis/xunit/System.Web.Tests.HttpUtilityTest~2FHtmlEncode_TextWriter(decoded:%20null,%20encoded:%20%5C%22%5C%22)
```text
Message :
System.ArgumentNullException : Value cannot be null.
Parameter name: value
Stack Trace :
   at System.Net.WebUtility.HtmlEncode(String value) in D:\j\workspace\windows-TGrou---74aa877a\src\System.Runtime.Extensions\src\System\Net\WebUtility.cs:line 42
   at System.Web.Util.HttpEncoder.HtmlEncode(String value, TextWriter output) in D:\j\workspace\windows-TGrou---74aa877a\src\System.Web.HttpUtility\src\System\Web\Util\HttpEncoder.cs:line 139
   at System.Web.HttpUtility.HtmlEncode(String s, TextWriter output) in D:\j\workspace\windows-TGrou---74aa877a\src\System.Web.HttpUtility\src\System\Web\HttpUtility.cs:line 140
   at System.Web.Tests.HttpUtilityTest.HtmlEncode_TextWriter(String decoded, String encoded) in D:\j\workspace\windows-TGrou---74aa877a\src\System.Web.HttpUtility\tests\HttpUtility\HttpUtilityTest.cs:line 273
```