@benjamin-hodgson 2018-03-09
@ahsonkhan Yep I know. I think I'm going to undo the `ArgumentNullException` change because it's a breaking change (see [my earlier comment](https://github.com/dotnet/corefx/pull/27250#issuecomment-371223676)). I haven't had the time yet though, will try and get to it over the weekend.