@benjamin-hodgson 2018-03-14
Apologies for the delay. I've rebased, and erased the commits with the changed `ArgumentNullException` behaviour