@danmosemsft 2018-03-14
Can you please share the latest perf data? Are we still 10-20% slower?