@stephentoub 2018-05-17
With the changes made in https://github.com/dotnet/corefx/pull/29402, is this PR still relevant, or are parts of it?  Assuming it is. it'll need to be rebased, and then it'll be good to see perf data.