@benjamin-hodgson 2018-05-18
@stephentoub I think it is still relevant, because the two PRs address different classes of allocations. I'll rebase this branch and get you some benchmarks next week