@benjamin-hodgson 2018-05-25
I measured `HtmlEncode` and `HtmlDecode`, with an input string of 100k `<`s and 100k `&lt;`s respectively.

```
         Method | Categories |     Mean |     Error |    StdDev | Scaled | ScaledSD |    Gen 0 |    Gen 1 |    Gen 2 |  Allocated |
--------------- |----------- |---------:|----------:|----------:|-------:|---------:|---------:|---------:|---------:|-----------:|
 HtmlEncode_Old |     Encode | 1.143 ms | 0.0227 ms | 0.0360 ms |   1.00 |     0.00 | 343.7500 | 330.0781 | 248.0469 | 1573.09 KB |
 HtmlEncode_New |     Encode | 2.319 ms | 0.0340 ms | 0.0318 ms |   2.03 |     0.07 | 281.2500 | 281.2500 | 281.2500 | 1109.29 KB |
                |            |          |           |           |        |          |          |          |          |            |
 HtmlDecode_Old |     Decode | 5.274 ms | 0.0712 ms | 0.0666 ms |   1.00 |     0.00 | 242.1875 | 242.1875 | 242.1875 |  976.73 KB |
 HtmlDecode_New |     Decode | 4.924 ms | 0.0305 ms | 0.0255 ms |   0.93 |     0.01 |  54.6875 |  54.6875 |  54.6875 |  203.34 KB |
```

So allocations are down across the board and `HtmlDecode` is about the same speed, but `HtmlEncode` is significantly slower. (This is using the Core 2.1 release candidate on OSX.)

I think this is down to the fact that `HtmlEncode` appends a few characters at a time to the output, whereas `HtmlDecode` only appends one character at a time in this benchmark. `ValueStringBuilder` seems to be comparatively slow in the former case. I opened a separate issue to look into this: #29921