@danmosemsft 2018-05-25
@benjamin-hodgson if you follow the "details" links you will see its' an unrelated (flaky) test failure. You can rerun like this:
@dotnet-bot test Windows x64 Debug Build please