@danmosemsft 2018-05-25
@stephentoub any other comments