@stephentoub 2018-05-25
> any other comments

I wouldn't want to merge this until the throughput regression is addressed.  It's great that allocation is reduced, but not at the expense of a 100% regression in throughput.