@ViktorHofer 2018-06-03
I benchmarked with these commits. See results below. Benchmark file: https://gist.github.com/ViktorHofer/d296de07c6178e586a0fb4f3f343e440

## Before

|                    Method |     Mean |     Error |    StdDev |       Gen 0 | Allocated |
|-------------------------- |---------:|----------:|----------:|------------:|----------:|
|              EncodeString | 76.35 ms | 0.6652 ms | 0.5897 ms | 100125.0000 |  200.5 MB |
| EncodeStringAndTextWriter | 77.63 ms | 1.2074 ms | 1.1294 ms | 100125.0000 |  200.5 MB |
|              DecodeString | 78.74 ms | 0.5254 ms | 0.4658 ms |   8937.5000 |  17.91 MB |
| DecodeStringAndTextWriter | 78.73 ms | 0.6507 ms | 0.5768 ms |   8937.5000 |  17.91 MB |

## After

|                    Method |     Mean |     Error |    StdDev |      Gen 0 |  Allocated |
|-------------------------- |---------:|----------:|----------:|-----------:|-----------:|
|              EncodeString | 79.47 ms | 1.8111 ms | 2.2243 ms | 36375.0000 | 76480000 B |
| EncodeStringAndTextWriter | 76.84 ms | 0.7241 ms | 0.5653 ms |          - |        0 B |
|              DecodeString | 79.44 ms | 1.0369 ms | 0.9192 ms |  1562.5000 |  3296000 B |
| DecodeStringAndTextWriter | 79.58 ms | 1.1509 ms | 1.0203 ms |          - |        0 B |