@ViktorHofer 2018-06-03
@benjamin-hodgson I rebased onto master and removed the StringBuilderCache reference in S.R.Extensions after it was removed from all other places with yesterdays change.