@ViktorHofer 2018-06-03
I'm looking into further optimizations in the WebUtility class. Please stand by.