@benjamin-hodgson 2018-06-03
@ViktorHofer Ah thanks very much for looking into that. Those results look really good - is the improvement due to https://github.com/dotnet/coreclr/issues/18005 ? If so should #29921 be closed now?