@ViktorHofer 2018-06-03
With @jkotas suggestion of increasing the buffer size (13ms) and switching to stackalloc for small inputs (12ms) I was able to make it faster than it currently is in master.

## Before

|                    Method |     Mean |     Error |    StdDev |      Gen 0 | Allocated |
|-------------------------- |---------:|----------:|----------:|-----------:|----------:|
|              EncodeString | 51.74 ms | 0.8133 ms | 0.6791 ms | 19062.5000 |  38.15 MB |
| EncodeStringAndTextWriter | 52.68 ms | 1.0402 ms | 1.3156 ms | 19062.5000 |  38.15 MB |
|              DecodeString | 83.84 ms | 1.4484 ms | 1.3549 ms |  5812.5000 |  11.75 MB |
| DecodeStringAndTextWriter | 86.87 ms | 1.5921 ms | 1.4892 ms |  5812.5000 |  11.75 MB |

## After

|                    Method |     Mean |     Error |    StdDev |      Gen 0 |  Allocated |
|-------------------------- |---------:|----------:|----------:|-----------:|-----------:|
|              EncodeString | 44.29 ms | 0.5136 ms | 0.4289 ms | 16000.0000 | 33600000 B |
| EncodeStringAndTextWriter | 50.87 ms | 0.9949 ms | 0.9772 ms |          - |        0 B |
|              DecodeString | 90.22 ms | 1.7041 ms | 1.5940 ms |  5812.5000 | 12320000 B |
| DecodeStringAndTextWriter | 94.55 ms | 1.8487 ms | 1.8157 ms |          - |        0 B |

**Note: I haven't yet looked at Decode at all...**